Pressure Sensor and Touch Sensor
The pressure sensor is to detect the weight of the trays and plates and
it will send to the management via the dashboard to request someone to
clear it. Another will be the touch sensor and a display board. Once the
sensor is touched, the sensor will send it to the management and it will
send a housekeeper up to clean the house.